remove usage of apollo-testing-support symbols that pull MockServer
